requirements.txt
boto3==1.38.17
botocore==1.38.17
bpmn-cwp-verify @ file:///Users/tyklabacka/Desktop/spreadsheet_automation
cachetools==5.5.2
certifi==2025.4.26
charset-normalizer==3.4.2
dotenv==0.9.9
google-api-core==2.24.2
google-api-python-client==2.169.0
google-auth==2.40.1
google-auth-httplib2==0.2.0
google-oauth2-tool==0.0.3
googleapis-common-protos==1.70.0
httplib2==0.22.0
idna==3.10
jmespath==1.0.1
oauth2client==4.1.3
proto-plus==1.26.1
protobuf==6.31.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
requests==2.32.3
returns==0.25.0
rsa==4.9.1
s3transfer==0.12.0
six==1.17.0
typing_extensions==4.13.2
uritemplate==4.1.1
urllib3==2.4.0
